WheelsSwap/requirements.txt
pillow
opencv-python
numpy
aiogram==3.*
pydantic-settings
python-dotenv
pytest
httpx
fastapi
uvicorn
python-multipart
